Moved integration test data file
